catalog + product with api methods

+++ lib/constants/utils.dart
@@ -293,7 +293,7 @@ class Utils {
     }
   }
 
-  static String formatPrice(double? price) {
+  static String formatPrice(int? price) {
     if (price == null) return '-';
     final NumberFormat formatter =
         NumberFormat.currency(locale: 'ru_RU', symbol: '₽', decimalDigits: 0);
